requirements.txt
# Pin loosely so installs work on current Python (e.g. 3.11–3.14); use a venv.
torch>=2.2.0
torchvision>=0.17.0
Pillow>=9.0.0
Flask>=2.3.0
Werkzeug>=2.3.0
flask-cors>=4.0.0
